Split up the classes in __init__.py into seperate files
