+++ email_reply_parser/fragment.py
@@ -1,0 +1,24 @@
+class Fragment(object):
+    """ A Fragment is a part of
+        an Email Message, labeling each part.
+    """
+
+    def __init__(self, quoted, first_line, headers=False):
+        self.signature = False
+        self.headers = headers
+        self.hidden = False
+        self.quoted = quoted
+        self._content = None
+        self.lines = [first_line]
+
+    def finish(self):
+        """ Creates block of content with lines
+            belonging to fragment.
+        """
+        self.lines.reverse()
+        self._content = '\n'.join(self.lines)
+        self.lines = None
+
+    @property
+    def content(self):
+        return self._content.strip()
